feat!: merge `userstyles` subcommands into root `query` and `init`
