+++ src/models/shared.rs
@@ -1,5 +1,7 @@
 use serde::{Deserialize, Serialize};
 
+use super::{ports::Port, userstyles::Userstyle};
+
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(untagged)]
 pub enum StringOrStrings {
@@ -55,4 +57,28 @@ pub const CATEGORIES: [&str; 35] = [
 pub struct Maintainer {
 	pub name: Option<String>,
 	pub url: String,
+}
+
+impl From<Userstyle> for Port {
+	fn from(userstyle: Userstyle) -> Self {
+		Port {
+			name: match userstyle.name {
+				StringOrStrings::Single(s) => s,
+				StringOrStrings::Multiple(s) => s.join("/"),
+			},
+			categories: userstyle.categories,
+			upstreamed: Some(false),
+			platform: StringOrStrings::Single("agnostic".to_string()),
+			url: Some(match userstyle.readme.app_link {
+				StringOrStrings::Single(s) => s,
+				StringOrStrings::Multiple(s) => s[0].clone(),
+			}),
+			links: None,
+			icon: userstyle.icon,
+			color: userstyle.color,
+			alias: None,
+			current_maintainers: userstyle.current_maintainers,
+			past_maintainers: userstyle.past_maintainers,
+		}
+	}
 }
